Fixed some warnings and partially updated the code to Objective-C 2.0.

+++ MI_DIPShape.h
@@ -48,6 +48,6 @@
 
 @property NSString* name;
 
-- (instancetype) initWithNumberOfPins:(int)numPins;
+- (instancetype) initWithNumberOfPins:(NSInteger)numPins;
 
 @end
